--- src/pages/CreateTodoPage/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import TopTitle from '@/components/common/TopTitle';
-import InnerCon from '@/components/common/InnerCon';
-
-const CreateTodoPage = () => (
-  <InnerCon>
-    <TopTitle title="일정 생성" />
-  </InnerCon>
-);
-
-export default CreateTodoPage;
